requirements.txt
# requirements.txt
# platform: win-64

asttokens=2.2.1=pypi_0
attrs=22.1.0=pypi_0
azure-core=1.26.1=pypi_0
azure-identity=1.12.0=pypi_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
broadbean=0.11.0=pypi_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
bzip2=1.0.8=he774522_0
ca-certificates=2022.10.11=haa95532_0
cachetools=5.2.0=pypi_0
certifi=2022.12.7=py310haa95532_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.2=pypi_0
contextlib2=21.6.0=pypi_0
contourpy=1.0.5=py310h59b6b97_0
cryptography=38.0.4=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.4=pypi_0
decorator=5.1.1=pypi_0
docopt=0.6.2=pypi_0
entrypoints=0.4=pypi_0
executing=1.2.0=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
future=0.18.2=pypi_0
gettext=0.19.8.1=ha2e2712_1008
glib=2.72.1=h7755175_0
glib-tools=2.72.1=h7755175_0
google-api-core=2.11.0=pypi_0
google-auth=2.15.0=pypi_0
googleapis-common-protos=1.57.0=pypi_0
gst-plugins-base=1.20.3=h001b923_2
gstreamer=1.20.3=h6b5321d_2
h5netcdf=1.1.0=pypi_0
h5py=3.7.0=pypi_0
icu=70.1=h0e60522_0
idna=3.4=pypi_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.19.2=pypi_0
ipython=8.7.0=pypi_0
ipywidgets=8.0.3=pypi_0
jedi=0.18.2=pypi_0
jpeg=9e=h2bbff1b_0
jsonschema=4.17.3=pypi_0
jupyter-client=7.4.8=pypi_0
jupyter-core=5.1.0=pypi_0
jupyterlab-widgets=3.0.4=pypi_0
kiwisolver=1.4.2=py310hd77b12b_0
krb5=1.19.3=h1176d77_0
lazy-loader=0.1=pypi_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hf4acfee_0
libclang13=14.0.6=default_h8992b79_0
libdeflate=1.8=h2bbff1b_5
libffi=3.4.2=hd77b12b_6
libglib=2.72.1=h3be07f2_0
libiconv=1.16=h2bbff1b_2
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libtiff=4.4.0=h8a3f274_2
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=h2bbff1b_0
libwebp-base=1.2.4=h2bbff1b_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
libzlib=1.2.13=hcfcfb64_4
lz4-c=1.9.4=h2bbff1b_0
matplotlib-base=3.6.2=py310h1094b8e_0
matplotlib-inline=0.1.6=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
msal=1.20.0=pypi_0
msal-extensions=1.0.0=pypi_0
munkres=1.1.4=py_0
nest-asyncio=1.5.6=pypi_0
nidaqmx-python=0.6.4=pyhd8ed1ab_0
numpy=1.23.4=py310h60c9a35_0
numpy-base=1.23.4=py310h04254f7_0
opencensus=0.11.0=pypi_0
opencensus-context=0.1.3=pypi_0
opencensus-ext-azure=1.1.7=pypi_0
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1s=h2bbff1b_0
packaging=22.0=py310haa95532_0
pandas=1.5.2=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pypi_0
pillow=9.3.0=py310hdc2b20a_1
pip=22.3.1=py310haa95532_0
pipreqs=0.4.11=pypi_0
platformdirs=2.6.0=pypi_0
ply=3.11=py310haa95532_0
portalocker=2.6.0=pypi_0
prompt-toolkit=3.0.36=pypi_0
protobuf=4.21.12=pypi_0
psutil=5.9.4=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pypi_0
pygments=2.13.0=pypi_0
pyjwt=2.6.0=pypi_0
pyparsing=3.0.9=py310haa95532_0
pyqt=5.15.7=py310hbabf5d4_0
pyqt5-sip=12.11.0=py310h8a704f9_0
pyqtgraph=0.12.3=py310haa95532_0
pyreq=1.2.0=pypi_0
pyrsistent=0.19.2=pypi_0
python=3.10.8=h966fe2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.10=2_cp310
pytz=2022.7=pypi_0
pyvisa=1.12.0=pypi_0
pywin32=305=pypi_0
pyzmq=24.0.1=pypi_0
qcodes=0.36.0=pypi_0
qcodes-contrib-drivers=0.17.0=pypi_0
qt=5.15.4=h2d74725_0
qt-main=5.15.4=h467ea89_2
qt-webengine=5.15.4=h7325c83_3
requests=2.28.1=pypi_0
rsa=4.9=pypi_0
ruamel-yaml=0.17.21=pypi_0
ruamel-yaml-clib=0.2.7=pypi_0
schema=0.7.5=pypi_0
setuptools=65.5.0=py310haa95532_0
sip=6.6.2=py310hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.40.0=h2bbff1b_0
stack-data=0.6.2=pypi_0
tabulate=0.9.0=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=pypi_0
tornado=6.2=py310h2bbff1b_0
tqdm=4.64.1=pypi_0
traitlets=5.7.1=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022g=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
uncertainties=3.1.7=pypi_0
urllib3=1.26.13=pypi_0
vc=14.2=h21ff451_1
versioningit=2.1.0=pypi_0
vs2015_runtime=14.32.31332=h1d6e394_9
wcwidth=0.2.5=pypi_0
websockets=10.4=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.4=pypi_0
wincertstore=0.2=py310haa95532_2
wrapt=1.14.1=pypi_0
xarray=2022.12.0=pypi_0
xz=5.2.8=h8cc25b3_0
yarg=0.1.9=pypi_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h19a0ad4_0
